This is the php app for Open Shift
I wanted to simply type in a simple serverside program on Open Shift, no luck.  Now, I'm typing it into Git Hub and will have to clone it to desktop, and clone it again to Open Shift.  If you know how to shift a repo from Git Hub to Open Shift--let me know!